fix(pf4): scroll to top on switching to a different step
For instance, when there are two step that are too long for the
wizard container (e.g. too many fields) and are next to each other.
If you scroll down in one of them and switch to the other, the
location of your scroll bar won't reset to the top but will stay
at the bottom.

This is bad because the user expects the scroll bar to go back
to the top of the next step.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>