fix(manager): rebuild validation engine
Validation can now correctly evaluate fields with multiple registered validators.
Every validator is converted to promise, regardless if it is async or
sync validator. This allows us to properly evaluate fields with multiple
async validators. Sync validation order is preserved due to the
sequential nature of the code execution and immediate promise
resulution.

One downside is that we have run tests with setImmediate to wait for the
manager api state update.

Benefit is, that we can have more than one async validator and can it
does not have to be at the start of the validate array as it is right
now in the renderer.