fix(pf4): fix bug in wizard navigation toolbar
In Patternfly 4, the WizardNavItem component `step` prop is "the step
passed into the onNavItemClick callback".

http://v4-archive.patternfly.org/v4/components/wizard#wizardnavitem

In Patternfly 5, onNavItemClick no longer exists. Instead, there is an
onClick callback which expects the index of the step to be navigated to.

https://www.patternfly.org/components/wizard/#wizardnavitem

However, filtering the navSchema like so: `.filter((field) =>
field.primary)` results in substeps with field.primary===false being
dropped. Practically, this means the 2nd substep and those following it.
Therefore, the index provided by `.map((step, ind) => ...` is incorrect.

The correct index can still be found from `step.index`.